% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject_yaml.R
\name{inject_yaml_md_output}
\alias{inject_yaml_md_output}
\title{injects output yaml}
\usage{
inject_yaml_md_output(rmd, write = TRUE, ...)
}
\arguments{
\item{rmd}{rmarkdown document to change yaml}

\item{write}{logical whether to overwrite the existing rmd file or just return lines}
}
\value{
TRUE if writing or vector of output lines of write=F
}
\description{
injects output yaml
}

