% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject_yaml.R
\name{output_indices}
\alias{output_indices}
\title{output indices containing the ouput chunk}
\usage{
output_indices(lines)
}
\arguments{
\item{lines}{file lines}
}
\value{
beginning and end index for output
}
\description{
output indices containing the ouput chunk
}

